Minor refactorings
- Fix issue with loadExternalisedExamplesAndListUnloadableExamples
- Combining String with list of string
