+++ core/src/main/kotlin/io/specmatic/conversions/ExampleFromFile.kt
@@ -6,7 +6,6 @@ import io.specmatic.core.HttpResponse
 import io.specmatic.core.Result
 import io.specmatic.core.SpecmaticConfig
 import io.specmatic.core.examples.server.SchemaExample
-import io.specmatic.core.filters.ExampleFilterContext
 import io.specmatic.core.log.logger
 import io.specmatic.core.pattern.*
 import io.specmatic.core.value.EmptyString
@@ -33,10 +32,6 @@ class ExampleFromFile(private val scenarioStub: ScenarioStub, val file: File) {
     constructor(file: File, strictMode: Boolean = true): this(scenarioStub = ScenarioStub.readFromFile(file, strictMode), file = file)
     constructor(json: JSONObjectValue, file: File, strictMode: Boolean = true): this(scenarioStub = ScenarioStub.parse(json, strictMode), file = file)
     constructor(scenarioStub: ScenarioStub) : this(scenarioStub, File(scenarioStub.filePath.orEmpty()))
-
-    fun toFilterContext(): ExampleFilterContext {
-        return ExampleFilterContext(scenarioStub)
-    }
 
     fun toRow(specmaticConfig: SpecmaticConfig = SpecmaticConfig()): Row {
         logger.log("Loading test file ${this.expectationFilePath}")
